Build GitHub pages on commit to master
Set up a GitHub workflow to build the documentation on commit to master and
commit built output html pages to gh-pages branch so that if repository
configured to use that for GitHub pages they will be served on github.io.

Needs a personal access token named ghpagespat set up for the repository.